Fix initialization of buffers for processing.
This fixes a longstanding issue with seeking.

Only input buffers should be saved in the port queue, output buffers
should be sent directly to omx.

Before, the output buffers where saved in the port queue, so the output
thread was trying to process them. This is not a big issue for the
initial buffers because they are empty, but in the middle of a flush
they are not, so the buffers before the flush where pushed to the sink.