Cleanup flushing mechanism for seeking.
Flushes for seeking works differently in input and output ports.

In input ports a real omx flush command is issued, then wait for
completion.

In output ports the internal queue is flushed, but no flush command is
sent to omx. There's still the possibility that the output buffers
coming from omx are wrong (before the flush) so they need to be tainted
somehow.