basefilter: fix init race condition
This fixes an issue when changing the state to READY too fast. The
problem comes when there's a state change from PAUSED to READY while
pad_chain is in the middle of initializing OpenMAX; the two threads
compete to initialize and de-initialize OpenMAX at the same time.

This is a simplified version of a patch developed by Mark Nauwelaerts.

Ideally the initialization should move out of pad_chain to avoid these
issues.

Signed-off-by: Felipe Contreras <felipe.contreras@nokia.com>